refactor: Separate Qr utility functions from staff Qr validation
